fixes to get CI to run
