+++ TrackletAlgorithm/MemoryTemplateBinnedCM.h
@@ -228,6 +228,9 @@ class MemoryTemplateBinnedCM{
 
   static constexpr int getWidth() {return DataType::getWidth();}
 
+#endif
+
+#ifdef CMSSW_GIT_HASH
   std::string name_;
   void setName(std::string name) { name_ = name;}
   std::string const& getName() const { return name_;}
@@ -235,8 +238,18 @@ class MemoryTemplateBinnedCM{
   unsigned int iSector_;
   void setSector(unsigned int iS) { iSector_ = iS;}
   unsigned int getSector() const { return iSector_;}
-  
-#endif
+
+  std::string decodeToBits(unsigned int field, unsigned int size) const {
+    unsigned int valtmp = field;
+    std::string str = "";
+    for(unsigned int i=0; i< size; i++) {
+      str = ((valtmp & 1) ? "1" : "0") + str;
+      valtmp >>= 1;
+    }
+    return str;
+  }
+#endif 
+
   
 };
